Add pipeline module and CLI refactor
